+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Remove iframe baking from `BakeAutotitledNotes`
+* Define rex_link on `ElementBase`
+* Link to rex from iframe in `BakeIframes`
+
 ## [v1.5.0] - 2022-4-08
 
 * Add `BakeFootnotes` to ap physics